# openapi.yaml
...
openapi: "3.1.0"
...
info:
  title: "NomiFun Model Gateway Protocol v1"
  version: "1.0"
  summary: "Independent community-operated gateway contract and M0 synthetic mock"
  description: "NomiFun official does not operate gateway instances, sell API access or run paid services. Community partners independently operate instances. Clients configure the base URL at runtime. This document freezes control-plane fields and native passthrough boundaries; opaque native schemas are not exhaustive upstream validators. Unknown additive control fields are accepted. credentials cannot enter URLs except the sole inbound Gemini ?key compatibility on its two native routes; never use query credentials in desktop clients, examples, redirects or logs. Auth conflicts/malformed sources fail401. All responses require x-request-id and Cache-Control:no-store;429 also Retry-After. See docs/protocol/v1.md for semantic and affinity obligations."
  license:
    name: "Apache License 2.0"
    identifier: "Apache-2.0"
servers:
  -
    url: "/"
    description: "Runtime-configured deployment API root; no official operated instance or compiled-in commercial URL."
...
security:
  -
    BearerAuth: []
  -
    XAPIKeyAuth: []
  -
    XGoogAPIKeyAuth: []
paths:
...
  "/v1/messages/count_tokens":
    post:
      operationId: "countMessageTokens"
      tags:
        - "Native inference"
      summary: "Native anthropic passthrough"
      description: "Optional; advertised by /v1/messages/count_tokens in meta.optional_endpoints. Native payload validation remains native. No cross-protocol conversion. Unknown fields and stream deltas/signatures must survive production relay."
      parameters:
        -
          name: "anthropic-version"
          in: "header"
          required: false
          schema:
            type: "string"
            minLength: 1
          description: "Optional at gateway ingress. Preserve an explicit native version unchanged; when absent, M1 uses and declares the operator-configured default without version conversion."
        -
          name: "X-NomiFun-Session-ID"
          in: "header"
          required: false
          schema:
            type: "string"
            minLength: 1
          description: "Optional key-scoped cache-affinity identity. 24h idle lifetime; expired/lost bindings fail closed, requiring a new session ID."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NativeBody"
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
        "400":
          $ref: "#/components/responses/AnthropicFailure"
        "401":
          $ref: "#/components/responses/AnthropicFailure"
        "402":
          $ref: "#/components/responses/AnthropicFailure"
        "403":
          $ref: "#/components/responses/AnthropicFailure"
        "404":
          $ref: "#/components/responses/AnthropicFailure"
        "405":
          $ref: "#/components/responses/AnthropicFailure"
        "413":
          $ref: "#/components/responses/AnthropicFailure"
        "415":
          $ref: "#/components/responses/AnthropicFailure"
        "429":
          $ref: "#/components/responses/AnthropicRateLimited"
        "500":
          $ref: "#/components/responses/AnthropicFailure"
        "501":
          $ref: "#/components/responses/AnthropicFailure"
        "502":
          $ref: "#/components/responses/AnthropicFailure"
        "503":
          $ref: "#/components/responses/AnthropicFailure"
        "504":
          $ref: "#/components/responses/AnthropicFailure"
        default:
          $ref: "#/components/responses/AnthropicFailure"
...
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      description: "Instance API key; same key accepted by each header alternative. Conflicting/malformed/repeated sources reject401."
    XAPIKeyAuth:
      type: "apiKey"
      in: "header"
      name: "x-api-key"
      description: "Same instance API key accepted on every protected endpoint."
    XGoogAPIKeyAuth:
      type: "apiKey"
      in: "header"
      name: "x-goog-api-key"
      description: "Same instance API key accepted on every protected endpoint."
    GeminiQueryKeyAuth:
      type: "apiKey"
      in: "query"
      name: "key"
      description: "Sole legacy inbound URL-key exception on the two native Gemini routes. Use headers for desktop/examples. Query key must never appear in access logs or redirects; other routes reject key."
  headers:
    RequestID:
      required: true
      description: "Non-empty credential-free request identifier."
      schema:
        type: "string"
        minLength: 1
    NoStore:
      required: true
      description: "All responses must forbid caching of account/model/error data."
      schema:
        type: "string"
        const: "no-store"
    RetryAfter:
      required: true
      description: "Required on429. Seconds or HTTP-date per HTTP; mock emits seconds."
      schema:
        type: "string"
        minLength: 1
  responses:
...
    AnthropicFailure:
      description: "Native error preserved unchanged. Gateway-generated failures use AnthropicGatewayError; upstream errors need not contain gateway extensions."
      headers:
        x-request-id:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/NativeBody"
          x-generated-error-schema:
            $ref: "#/components/schemas/AnthropicGatewayError"
    AnthropicRateLimited:
      description: "HTTP 429: rate_limited or native upstream rate limit. Retry-After required; gateway error extension follows AnthropicGatewayError."
      headers:
        x-request-id:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/NativeBody"
          x-generated-error-schema:
            $ref: "#/components/schemas/AnthropicGatewayError"
...
  schemas:
...
    GatewayErrorCode:
      type: "string"
      enum:
        - "insufficient_balance"
        - "subscription_expired"
        - "model_not_in_plan"
        - "key_expired"
        - "rate_limited"
        - "invalid_api_key"
        - "invalid_request_error"
        - "invalid_previous_response_id"
        - "not_found"
        - "method_not_allowed"
        - "request_too_large"
        - "unsupported_media_type"
        - "model_not_found"
        - "unsupported_endpoint"
        - "stream_unavailable"
      description: "Gateway-generated error codes. Upstream native errors can have other codes and retain native structure."
...
    NativeBody:
      type: "object"
      additionalProperties: true
      description: "Opaque native JSON object. This schema intentionally does not enumerate or validate all upstream native fields. Preserve unknown fields, thinking/signatures, cache details and tools."
...
    AnthropicGatewayError:
      type: "object"
      required:
        - "type"
        - "error"
      properties:
        type:
          type: "string"
          const: "error"
        error:
          type: "object"
          required:
            - "type"
            - "message"
            - "code"
            - "purchase_url"
          properties:
            type:
              type: "string"
              minLength: 1
            message:
              type: "string"
              minLength: 1
            code:
              $ref: "#/components/schemas/GatewayErrorCode"
            purchase_url:
              anyOf:
                -
                  type: "string"
                  format: "uri"
                  pattern: "^https://[^/?#@]+(?:[/?#].*)?$"
                  description: "Absolute HTTPS URL without authority credentials; must not carry API keys."
                -
                  type: "null"
          additionalProperties: true
      additionalProperties: true
...